refactor: extract shared toast and clipboard utilities
Keep reusable clipboard and toast context wiring in shared utilities.
Update existing consumers and tests to use the new shared imports.
