+++ frontend/src/shared/Toast.tsx
@@ -1,27 +1,18 @@
 import {
-    createContext,
     useCallback,
-    useContext,
     useEffect,
     useMemo,
     useRef,
     useState,
 } from "react";
 import type { ReactNode } from "react";
-
-export type ToastType = "success" | "error" | "info";
+import { ToastContext, type ToastType } from "@/shared/toast-context";
 
 interface ToastItem {
     id: number;
     message: string;
     type: ToastType;
 }
-
-export interface ToastContextValue {
-    showToast: (message: string, type?: ToastType) => void;
-}
-
-const ToastContext = createContext<ToastContextValue | null>(null);
 
 export function ToastProvider({ children }: { children: ReactNode }) {
     const [toasts, setToasts] = useState<ToastItem[]>([]);
@@ -76,12 +67,4 @@ export function ToastProvider({ children }: { children: ReactNode }) {
             </div>
         </ToastContext.Provider>
     );
-}
-
-export function useToast(): ToastContextValue {
-    const ctx = useContext(ToastContext);
-    if (!ctx) {
-        throw new Error("useToast must be used inside ToastProvider");
-    }
-    return ctx;
 }
