+++ frontend/src/shared/toast-context.ts
@@ -1,0 +1,17 @@
+import { createContext, useContext } from "react";
+
+export type ToastType = "success" | "error" | "info";
+
+export interface ToastContextValue {
+    showToast: (message: string, type?: ToastType) => void;
+}
+
+export const ToastContext = createContext<ToastContextValue | null>(null);
+
+export function useToast(): ToastContextValue {
+    const ctx = useContext(ToastContext);
+    if (!ctx) {
+        throw new Error("useToast must be used inside ToastProvider");
+    }
+    return ctx;
+}
